pyproject.toml
[tool.poetry]
name = "categoraize"
version = "0.1.0"
description = "Система персонализированной классификации финансовых трат пользователей с использованием ML"
authors = ["MoneyCluster Team"]
readme = "README.md"
packages = [{include = "categoraize", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
sentence-transformers = "^2.7.0"
scikit-learn = "^1.5.0"
transformers = "^4.46.0"
huggingface-hub = "^0.24.0"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
pydantic = "^2.9.0"
pandas = "^2.2.0"
numpy = "^2.1.0"
pyyaml = "^6.0.1"
joblib = "^1.4.0"
tqdm = "^4.66.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
pre-commit = "^3.7.0"

[tool.poetry.scripts]
categoraize-train = "categoraize.train:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "ISC", "ICN", "T20", "SIM", "ARG", "PTH", "PL", "TRY", "RUF"]
ignore = ["RUF002"]  # Игнорируем предупреждения о кириллице в docstrings

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/categoraize --cov-report=term-missing --cov-report=html"

